feat: improve search functionality in blog
Adjusted the weight attributed to 'title' and 'body' in Search Vector and amended the ranking filter threshold in Search Rank. The aim is to enhance the blog's search functionality and provide better result ranking. Future improvements in search ranking will utilize the TrigramExtension, for which a new migration has been added.